+++ galacteek/ipfs/ipfsops/nscache.py
@@ -2,7 +2,6 @@ import json
 import aiofiles
 import asyncio
 import time
-import traceback
 from pathlib import Path
 
 from galacteek import log
@@ -54,7 +53,6 @@ class IPNSCache:
             if not jsonSchemaValidate(cache, nsCacheSchema):
                 raise Exception('Invalid NS cache schema')
         except Exception as err:
-            traceback.print_exc()
             log.debug(f'Error loading NS cache: {err}')
         else:
             log.warning(f'IPNS cache: loaded from {self.nsCachePath}')