+++ galacteek/ui/icapsules/__init__.py
@@ -2,7 +2,6 @@ from rdflib import URIRef
 
 from PyQt5.QtWidgets import QWidget
 from PyQt5.QtWidgets import QVBoxLayout
-from PyQt5.QtWidgets import QToolButton
 from PyQt5.QtWidgets import QListView
 from PyQt5.QtWidgets import QLabel
 
@@ -14,7 +13,6 @@ from galacteek.core import runningApp
 
 from galacteek.core.models.sparql.icapsules import ICapsulesSparQLModel
 
-from ..helpers import getIcon
 from ..helpers import messageBoxAsync
 
 
@@ -69,19 +67,4 @@ class ICapsulesManagerWidget(QWidget):
         self.queryDapps()
 
     def queryDapps(self):
-        print(self.icapsuledb.querier.qAllDappManifests)
         self.model.graphQuery(self.icapsuledb.querier.qAllDappManifests)
-
-
-class ICapsulesManagerButton(QToolButton):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-        self.app = runningApp()
-        self.clicked.connect(self.onOpen)
-
-        self.setIcon(getIcon('capsules/icapsule-green.png'))
-
-    def onOpen(self):
-        self.app.mainWindow.wspaceDapps.refresh()
-        self.app.mainWindow.wspaceDapps.wsSwitch()